- Добивание загрузки фотографий; - Метод photos.save; - Исправление удаления фотографий, а также удаления самого файла;

+++ src/handlers/photos/remove.ts
@@ -1,5 +1,9 @@
 import { ICallPropsPrivate, PrivateMethodAPI } from '../method';
 import { IApiParams } from '../../types/api';
+import { existsSync, unlinkSync } from 'fs';
+import { IPhotoRaw } from '../../types/photo';
+import { ApiError, ErrorCode } from '../../error';
+import { getFullFilePath } from '../../uploader/utils/get-full-file-path';
 
 type IParams = {
     photoId: number;
@@ -16,13 +20,29 @@ export default class PhotosRemove extends PrivateMethodAPI<IParams, boolean> {
         return { photoId };
     }
 
-    // TODO доделать удаление самих файлов
+
     protected async perform({ photoId }: IParams, { database, session }: ICallPropsPrivate): Promise<boolean> {
+        const photos = await database.select<IPhotoRaw>(
+            'select `path`, `photo200`, `photoMax` from `photo` where `photoId` = ? and `ownerId` = ?',
+            [photoId, session.userId],
+        );
+
+        if (photos.length !== 1) {
+            throw new ApiError(ErrorCode.ACCESS_DENIED, 'Access denied');
+        }
+
+        const photo = photos[0];
+
+        [photo.photo200, photo.photoMax]
+            .map(name => getFullFilePath(photo.path!, name))
+            .filter(path => existsSync(path))
+            .forEach(path => unlinkSync(path));
+
         const result = await database.apply(
             'delete from `photo` where `photoId` = ? and `ownerId` = ?',
             [photoId, session.userId],
         );
 
-        return result.affectedRows === 0;
+        return result.affectedRows > 0;
     }
 }
